Refactor SSH auth for greater extensibility
This introduces a dispatcher and splits the authorized keys command and
the git.sr.ht shell into two binaries. The dispatcher is now to be set
as the authorized keys function, and some configuration allows you to
configure how it dispatches to the authentication agent.

This will be necessary for implementing man.sr.ht's git support.

This change requires manual intervention to apply.