Speed up finding latest commits
Walking through the repo is very slow on larger repositories, this makes
loading the index page for large repos slow, for example it takes tens
of seconds to show https://git.sr.ht/~sircmpwn/linux

Since finding parents of commits is fast, and since the last 3 commits
(including the given one) must be within the parents and grandparents
of the given commit, the same task can be performed in much less time.

This implementation is instantaneous even for the Linux repo.