Support installing into a virtual environment
It turns out that `virtualenv` has a custom (and non quite working)
implementation of the `site` module, and it doesn't have the
`getsitepackages` method... so since we already need the `$SRHT_PATH`
environment variable, we can use that instead to locate the core `srht`
package.

We also give an error so the user knows what's going on.

Finally, we support situations where Python installs packages in the
`USER_BASE` directory.