Filter duplicate refs in gitsrht-update-hook
If multiple refs that point to the same commit are pushed with one
command, the girsrht-update-hook currently triggers multiple builds.  To
fix this, we check the object IDs of the Git references and skip
duplicates.  (We have to duplicate the checks as git_repo.get(ref)
returns a Git object and git_repo.lookup_reference(ref).target is an
object ID.)