Include funcname in hunk header in commit view if available
pygit2 unfortunately doesn't make the function name output directly available
as a property on the hunk like it does for the line info, but we can extract it
from the raw hunk header by splitting on the '@@'.  The function name will be
the last element and will be terminated by a '\n', so we strip that out. (If
there is no function name the string will only be the '\n')