Attach annotations to a specific commit sha
See git.sr.ht#267

When a new commit is pushed which updates some blobs, but not others,
cross-references between them will be broken.

We could technically associate this with the top-level tree sha, but
this would be wrong for subtrees and it's not obvious how to avoid this.
Since one commit is guaranteed to be associated with a toplevel tree,
attaching it to the commit solves this without much compromise.