Add git-daemon-export-ok to world readable repos
This is an extra security measure, not a replacement to the /authorize
method. In order to clone an HTTP(S) repo, the corresponding /authorize
request must return 200, and the git-daemon-export-ok must exist inside
the repository directory. The latter only will be required when the
GIT_HTTP_EXPORT_ALL environment variable is unset.