Set receive.denyDeleteCurrent=ignore for old and newly created repositories
By default, git disallows removing the current default branch, with a
large error notice like this:
-- >8 --
remote: error: By default, deleting the current branch is denied, because the next
remote: 'git clone' won't result in any file checked out, causing confusion.
remote:
remote: You can set 'receive.denyDeleteCurrent' configuration variable to
remote: 'warn' or 'ignore' in the remote repository to allow deleting the
remote: current branch, with or without a warning message.
remote:
remote: To squelch this message, you can set it to 'refuse'.
-- >8 --

However, we handle this ourselves in the post-update hook, and will only
let the default branch dangle if there are no branches (the repository
is empty), which is equivalent to the "fresh repository" state.

This also means that *all current repositories need to be migrated*
to the new config to be able to take advantage of the new semantics.