Fix UTF-8 replacement chars in git send-email
git-format-patch doesn't set the charset param of the Content-Type
header field. The default value is "ascii", which forbids UTF-8.
However git-format-patch message bodies may include UTF-8. This seems
to be deliberate to make the output of git-format-patch more readable
for humans.

get_content() tries to decode the body according to the charset
parameter of the Content-Type header field. Instead of avoiding to
use get_content() as a workaround (it's easy to forget about it and
step on a landmine), fixup the Content-Type header field to include
the proper charset param.

Closes https://todo.sr.ht/~sircmpwn/git.sr.ht/327