Allow "path" ref in log and tree URLs
This change updates both the <owner>/<repo>/log and
<owner>/<repo>/tree URLs by including a sentinel value between the two
possible "path"-like parts of the URL, the ref and the path.

Without a sentinel value it is not possible to use the two path
variables due to the greedy nature of the last instance.

Case in point, something like:

@route(<path:first>/<path:second>)
def example(first, second):
    return f'first: {first}    second: {second}'

GET debian/0.1.0-2/some/file

first: debian    second: 0.1.0-2/some/file