Fix logic for displaying "next" link on log page
Fetch one more commit from get_log and only display a "next" link if it
exists.  The current logic checks whether the last commit in the list
has a parent.  This doesn't work if the log is shown for a path, because
when the first commit for the path is not the initial commit it has a
parent, but there are no more log entries for the path.

The parameter name for the template is copied from hg.sr.ht which solves
this problem with an additional return parameter from get_log.