Fix link from commit to view commit in log
If a commit is old enough, it won't appear on the first page of the
git log, which means that the anchor in the URL won't work.

This change makes it so that the link always goes to a log page where
the commit is visible.

Note: This same change will fix the same issue for links on the log
page.