Allow uploading multiple artifacts at once
Useful to upload both the tarball and the PGP signature at once,
for instance.

For some reason getlist returns a non-empty list even if the user
selects no file. If the user selects one or multiple files, it
behaves as expected. That's why len(file_list) isn't used for
validation.