add whitespace back to pre, body on blob page
This is a partial revert of fd8cca2e. It fixes a bug where, for example, viewing
a README file on a blob page results in no bottom margin for <pre>s at all, and
no bottom padding on the body. Adding bottom padding to the body also improves
the experience on mobile.

Signed-off-by: Vlad-Stefan Harbuz <vlad@vladh.net>