api: micro-second resolution for repo pagination
Lists of repositories are sorted and paginated based on their "updated"
timestamp. However, fractional seconds are discarded. This can cause
issues with repositories that have been updated in the same second (e.g.
by running a script).

This commit leverages the micro-second resolution of the timestamp in
postgres to make this case not impossible, yet highly unlikely.