Stop using scmsrht.BaseScmOAuthService
Currently, both the git.sr.ht and hg.sr.ht OAuth implementation is based
on `scmsrht.BaseScmOAuthService`. The main value that class provides is
handling of SSH keys in a DB table for that service (including webhooks
to update them). However, this is exactly what we want to get rid of.
In fact this table is already never read - all SSH key reads already go
through meta.sr.ht.

Since this was the main feature of this base class, simply stop using it
and base the service on `srht.oauth.AbstractOAuthService`, like all
non-SCM-services do. Once the same has been done for hg.sr.ht, the code
in scm.sr.ht can be dropped.

Note that this commit will cause the (still active on meta.sr.ht) SSH
key update webhooks to 404. This was deemed acceptable. If so desired,
the hooks can be cleaned out of the database by running the following
query against meta.sr.ht:

    DELETE FROM user_webhook_subscription
    WHERE url = 'https://git.sr.ht/webhook/notify/keys'