Produce stable *.patch file downloads
There are two sources of instability with this endpoint:

- the version number for git itself is added by default as a trailer,
  and changes on a frequent basis whenever the server upgrades git

- the "index xxxxxxx..xxxxxxx 10064" line in the diff header contains a
  reference to git blobs, and by default git blobs are abbreviated as
  much as possible in order to produce a "short and succinct"
  description, but the cutoff changes over time as the repo gets larger

Fix both of these issues, by passing the relevant options whenever
running `git format-patch`. Although passing `--full-index` for the
latter case might ordinarily be considered a breaking change, the
checksum of the patch file contents was always completely unreliable due
to the former case, so this is our chance to get it correct for all
time.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>