api: handle all errors in Tree.Entry()
The FindEntry() function can return a similar "not found" error,
depending on the path passed in and circumstances. Currently, this error
goes unhandled, causing a nil derefence in the return statement. Add
this additional error type to the check to prevent the panic.

At least in theory, FindEntry() could also return completely different
errors from deep within go-git, so add an explicit panic in case this
ever happens, so we at least know what's happening.