update-hook: remove pointless statement preparation
fetchInfoForPush currently does two round-trips to the database to find
the owner's username and OAuth token: one to prepare the query, and one
to actually execute it.

Since the process is short-lived and prepared statements are cleaned by
the server upon session end, we get no benefit from preparing the
statement, but pay the code of going twice to the database.

This patch removes the preparation step.

Signed-off-by: Simon Martin <simon@nasilyan.com>