shell: use sourcehut-ssh/shell to parse arguments
This makes git.sr.ht-shell use the command line parsing logic provided
by sourcehut-ssh (as this is also where the command line arguments are
generated).

It also slightly reduces the amount of logs produced for the happy path,
as git.sr.ht-shell is a pretty notorious logger. It certainly is so for
valid reasons, but I think this change should still strike a decent
balance between only logging the most important bits in the happy path
and the relevant details on error.