githook: look at newHash, not oldHash
Since hub.sr.ht's 1678dec5 commit, the very first push to a git
repository linked to a project is not reflected anymore in hub, because
the payload of its associated webhook call to hub has no "new" attribute
and the update is ignored in hubsrht/blueprints/webhooks.py#L105.

We did not see the issue before because hub did not rely on "new" to
generate its associated event.

Fixes: efd03947 ("API, update-hook: deliver pre-receive/post-receive hooks")
Signed-off-by: Simon Martin <simon@nasilyan.com>